+++ super_repo/__init__.py
@@ -4,7 +4,7 @@
 
 Examples for the calculater functionality to demonstrate TDD.
 
-SPDX-FileCopyrightText: Jonas Huber (jh-RLI) © Reiner Lemoine Institut
-SPDX-FileCopyrightText: Ludwig Hülk (Ludee) © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Jonas Huber <@jh-RLI> © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Ludwig Hülk <@Ludee> © Reiner Lemoine Institut
 SPDX-License-Identifier: MIT
 """