+++ super_repo/example_calculator.py
@@ -4,8 +4,8 @@
 
 Examples for the calculater functionality to demonstrate TDD.
 
-SPDX-FileCopyrightText: Jonas Huber (jh-RLI) © Reiner Lemoine Institut
-SPDX-FileCopyrightText: Ludwig Hülk (Ludee) © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Jonas Huber <@jh-RLI> © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Ludwig Hülk <@Ludee> © Reiner Lemoine Institut
 SPDX-License-Identifier: MIT
 """