+++ tests/test_example.py
@@ -5,7 +5,7 @@
 Example implementation of common calculater functionality to
 demonstrate TDD.
 
-SPDX-FileCopyrightText: Jonas Huber (jh-RLI) © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Jonas Huber <@jh-RLI> © Reiner Lemoine Institut
 SPDX-License-Identifier: MIT
 """